+++ src/gateways/Gateway.php
@@ -143,7 +143,7 @@ class Gateway extends BaseGateway
     /**
      * string The Stripe API version to use.
      */
-    const STRIPE_API_VERSION = '2018-02-06';
+    const STRIPE_API_VERSION = '2018-07-27';
 
     // Properties
     // =========================================================================
@@ -506,27 +506,40 @@ class Gateway extends BaseGateway
     public function getSubscriptionPlans(): array
     {
         /** @var Collection $plans */
-        $plans = StripePlan::all();
+        $plans = StripePlan::all([
+            'limit' => 100,
+        ]);
         $output = [];
 
+        $planProductMap = [];
         $planList = [];
+
         if (\count($plans->data)) {
             foreach ($plans->data as $plan) {
                 $plan = $plan->jsonSerialize();
-                $planList[$plan['product']] = $plan['id'];
+                $planProductMap[$plan['id']] = $plan['product'];
+                $planList[] = $plan;
             }
 
             /** @var Collection $products */
             $products = StripeProduct::all([
                 'limit' => 100,
-                'ids' => array_keys($planList)
+                'ids' => array_values($planProductMap),
             ]);
+
+            $productList = [];
 
             if (\count($products->data)) {
                 foreach ($products->data as $product) {
                     $product = $product->jsonSerialize();
-                    $output[] = ['name' => $product['name'], 'reference' => $planList[$product['id']]];
+                    $productList[$product['id']] = $product;
                 }
+            }
+
+            foreach ($planList as $plan) {
+                $productName = $productList[$plan['product']]['name'];
+                $planName = null !== $plan['nickname'] ? ' ('.$plan['nickname'].')' : '';
+                $output[] = ['name' => $productName.$planName, 'reference' => $plan['id']];
             }
         }
 
@@ -1068,8 +1081,7 @@ class Gateway extends BaseGateway
             return;
         }
 
-        $childTransaction = Commerce::getInstance()->getTransactions()->createTransaction(null, $transaction);
-        $childTransaction->type = $transaction->type;
+        $childTransaction = Commerce::getInstance()->getTransactions()->createTransaction(null,     $transaction);
         $childTransaction->reference = $data['id'];
 
         try {
@@ -1179,7 +1191,7 @@ class Gateway extends BaseGateway
 
         // Find the relevant line item and update subscription end date
         foreach ($lineItems as $lineItem) {
-            if ($lineItem['id'] === $subscriptionReference) {
+            if (!empty($lineItem['subscription']) && $lineItem['subscription'] === $subscriptionReference) {
                 $payment = $this->_createSubscriptionPayment($invoice->invoiceData, $currency);
                 Commerce::getInstance()->getSubscriptions()->receivePayment($subscription, $payment, DateTimeHelper::toDateTime($lineItem['period']['end']));